get rid of driver.util
